fix
+ fix Charge
+ fix Expense
+ fix Warehouse
+ fix Sale

+++ src/main/java/app/service/ExpenseItemsService.java
@@ -1,12 +1,15 @@
 package app.service;
 
+import app.entity.Charge;
 import app.entity.ExpenseItem;
+import app.exceptions.ExpenseItemNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Service;
 import app.repository.ExpenseItemRepository;
 
 import javax.transaction.Transactional;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -20,15 +23,26 @@ public class ExpenseItemsService {
     @Transactional
     @Modifying
     public void deleteById(Integer id){
+        if(!repository.isExistsById(id)){
+            throw new ExpenseItemNotFoundException("Expense item not found by ID");
+        }
         repository.deleteById(id);
     }
 
     public ExpenseItem save(ExpenseItem expenseItem){
+        if(expenseItem.getName() == null){
+            throw new IllegalArgumentException("Name can`t be empty!");
+        }
         return repository.save(expenseItem);
     }
 
-    public Optional<ExpenseItem> findById(Integer id){
-        return repository.findById(id);
+    public ExpenseItem findById(Integer id){
+        Optional<ExpenseItem> item =  repository.findById(id);
+        if(item.isPresent()){
+            return item.get();
+        }else{
+            throw new ExpenseItemNotFoundException("Expense item not found by ID");
+        }
     }
 
     public boolean isExistsName(String name){
@@ -39,16 +53,16 @@ public class ExpenseItemsService {
         return repository.isExistsById(id);
     }
 
-    @Transactional
-    @Modifying
-    public void deleteByName(String name){
-        repository.deleteByName(name);
-    }
 
-    public Optional<ExpenseItem> findByName(String name){
-        return repository.findByName(name);
+    public ExpenseItem findByName(String name){
+        Optional<ExpenseItem> item =  repository.findByName(name);
+        if(item.isPresent()){
+            return item.get();
+        }else{
+            throw new ExpenseItemNotFoundException("Expense item not found by Name");
+        }
     }
-
+    public List<ExpenseItem> findAll(){return(List<ExpenseItem>) repository.findAll();}
 
 
 }
